+++ chat_toolkit/common/__init__.py
@@ -1,15 +1,14 @@
 from .constants import TMP_DIR
 from .custom_types import StartingPromptsType
-from .utils import (
-    RecordingEndedWithKeyboardSignal,
-    set_openai_api_key,
-    temporary_file,
-)
+from .exceptions import SpeakingRateError
+from .orchestrator import Orchestrator
+from .utils import set_openai_api_key, temporary_file
 
 __all__ = (
     "set_openai_api_key",
     "temporary_file",
-    "RecordingEndedWithKeyboardSignal",
+    "Orchestrator",
+    "SpeakingRateError",
     "StartingPromptsType",
     "TMP_DIR",
 )