+++ chat_toolkit/components/__init__.py
@@ -5,6 +5,10 @@ from .speech_to_text.openai_speech_to_text import OpenAISpeechToText
 from .speech_to_text.speech_to_text_component_base import (
     SpeechToTextComponentBase,
 )
+from .text_to_speech.pyttsx3_text_to_speech import Pyttsx3TextToSpeech
+from .text_to_speech.text_to_speech_component_base import (
+    TextToSpeechComponentBase,
+)
 
 __all__ = (
     "ChatbotComponentBase",
@@ -12,5 +16,7 @@ __all__ = (
     "CostEstimatorBase",
     "OpenAIChatBot",
     "OpenAISpeechToText",
+    "Pyttsx3TextToSpeech",
     "SpeechToTextComponentBase",
+    "TextToSpeechComponentBase",
 )