+++ chat_toolkit/components/component_base.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from typing import Optional
 
 
 class CostEstimatorBase(ABC):
@@ -19,7 +20,7 @@ class CostEstimatorBase(ABC):
 
     @property
     @abstractmethod
-    def cost_estimate_data(self) -> tuple[float, dict]:
+    def _cost_estimate_data(self) -> tuple[float, dict]:
         """
         Abstract property that returns the estimated cost in of an object's
         usage so far and any relevant metadata available.
@@ -28,13 +29,23 @@ class CostEstimatorBase(ABC):
         """
         pass
 
+    @property
+    def cost_estimate_data(self) -> tuple[float, dict]:
+        """
+        Takes private method and adds any standardized data to the metadata
+        before returning
+        """
+        cost_estimate, metadata = self._cost_estimate_data
+        metadata["pricing_rate"] = self._pricing_rate
+        return cost_estimate, metadata
+
 
 class ComponentBase(CostEstimatorBase, ABC):
     """
     Used to create components in a standardized manner.
     """
 
-    def __init__(self, model: str, **kwargs):
+    def __init__(self, model: Optional[str], **kwargs):
         """
         Instantiates component.