+++ test_suite/unit/test_main.py
@@ -1,3 +1,4 @@
+from typing import Optional
 from unittest.mock import Mock
 
 import pytest
@@ -13,9 +14,14 @@ from chat_toolkit.__main__ import COMPONENTS, main
     "speech_to_text_model",
     list(COMPONENTS["speech_to_text"].keys()) + [None],
 )
+@pytest.mark.parametrize(
+    "text_to_speech_model",
+    list(COMPONENTS["text_to_speech"].keys()) + [None],
+)
 def test_main(
     chatbot_model: str,
-    speech_to_text_model: str,
+    speech_to_text_model: Optional[str],
+    text_to_speech_model: Optional[str],
     monkeypatch: pytest.MonkeyPatch,
     patched_openai_speech_to_text: None,
 ) -> None:
@@ -24,11 +30,11 @@ def test_main(
     dynamically.
     """
     monkeypatch.setattr(
-        "chat_toolkit.orchestrators.orchestrator_base.OrchestratorBase"
+        "chat_toolkit.common.orchestrator.Orchestrator"
         ".terminal_conversation",
         Mock(),
     )
     try:
-        main(chatbot_model, speech_to_text_model)
+        main(chatbot_model, speech_to_text_model, text_to_speech_model)
     except ImportError as e:
         raise AssertionError from e