+++ test_suite/unit/test_openai_speech_to_text.py
@@ -4,19 +4,15 @@ from unittest.mock import Mock
 
 import pytest
 
-from chat_toolkit.common.utils import (
-    RecordingEndedWithKeyboardSignal,
-    temporary_file,
-)
+from chat_toolkit.common.utils import temporary_file
 from test_suite.unit.conftest import (
-    CHATBOT_MODEL_TYPES,
+    SPEECH_TO_TEXT_MODEL_TYPES,
     TEST_TEXT,
     OpenAISpeechToTextFactoryType,
-    TypeMatcher,
 )
 
 
-@pytest.mark.parametrize("model", CHATBOT_MODEL_TYPES)
+@pytest.mark.parametrize("model", SPEECH_TO_TEXT_MODEL_TYPES)
 @pytest.mark.parametrize("pricing_rate", [0.1, 0.05, 0.002, 1.0, 0.0])
 @pytest.mark.parametrize("device", [0])
 def test_init(
@@ -31,6 +27,7 @@ def test_init(
     speech_to_text = patched_openai_speech_to_text_factory(
         model, pricing_rate, device
     )
+    assert speech_to_text
     assert speech_to_text._model == model
     assert speech_to_text._pricing_rate == pricing_rate
     assert speech_to_text.seconds_transcribed == 0
@@ -38,37 +35,7 @@ def test_init(
     assert isinstance(speech_to_text.tmp_file_directory, Path)
 
 
-@pytest.mark.parametrize("model", CHATBOT_MODEL_TYPES)
-def test_record_and_transcribe_happy(
-    patched_openai_speech_to_text_factory: OpenAISpeechToTextFactoryType,
-    model: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """
-    Test that no exception is raised when user interrupts recording with a
-    KeyboardInterrupt.
-    """
-    speech_to_text = patched_openai_speech_to_text_factory(model)
-    monkeypatch.setattr(
-        speech_to_text,
-        "record_unspecified_length_audio",
-        Mock(
-            side_effect=RecordingEndedWithKeyboardSignal(
-                "Mock User Interruption"
-            )
-        ),
-    )
-    try:
-        speech_to_text.record_and_transcribe()
-    except RecordingEndedWithKeyboardSignal as e:
-        raise AssertionError from e
-
-    speech_to_text.record_unspecified_length_audio.assert_called_once_with(  # type: ignore  # noqa: E501
-        TypeMatcher(str)
-    )
-
-
-@pytest.mark.parametrize("model", CHATBOT_MODEL_TYPES)
+@pytest.mark.parametrize("model", SPEECH_TO_TEXT_MODEL_TYPES)
 def test_record_and_transcribe_sad(
     patched_openai_speech_to_text_factory: OpenAISpeechToTextFactoryType,
     model: str,
@@ -78,16 +45,17 @@ def test_record_and_transcribe_sad(
     Test that an exception is raised when another error is raised.
     """
     speech_to_text = patched_openai_speech_to_text_factory(model)
+    assert speech_to_text
     monkeypatch.setattr(
         speech_to_text,
         "record_unspecified_length_audio",
         Mock(side_effect=Exception("Mock Error")),
     )
     with pytest.raises(Exception, match="Mock Error"):
-        speech_to_text.record_and_transcribe()
+        speech_to_text.transcribe_speech()
 
 
-@pytest.mark.parametrize("model", CHATBOT_MODEL_TYPES)
+@pytest.mark.parametrize("model", SPEECH_TO_TEXT_MODEL_TYPES)
 @pytest.mark.parametrize("pricing_rate", [0.1, 0.05, 0.002, 1.0, 0.0])
 @pytest.mark.parametrize("seconds", [1000, 0, 5])
 def test_cost_estimate_data(
@@ -100,15 +68,19 @@ def test_cost_estimate_data(
     Test that cost estimate is calculated correctly.
     """
     speech_to_text = patched_openai_speech_to_text_factory(model, pricing_rate)
+    assert speech_to_text
     speech_to_text._seconds_transcribed = seconds
 
     cost_estimate, metadata = speech_to_text.cost_estimate_data
     assert cost_estimate == seconds / 60 * pricing_rate
-    assert metadata == {"seconds_transcribed": seconds}
+    assert metadata == {
+        "seconds_transcribed": seconds,
+        "pricing_rate": speech_to_text._pricing_rate,
+    }
     assert speech_to_text.seconds_transcribed == seconds
 
 
-@pytest.mark.parametrize("model", CHATBOT_MODEL_TYPES)
+@pytest.mark.parametrize("model", SPEECH_TO_TEXT_MODEL_TYPES)
 def test_transcribe(
     patched_openai_speech_to_text_factory: OpenAISpeechToTextFactoryType,
     model: str,
@@ -119,6 +91,7 @@ def test_transcribe(
     call.
     """
     speech_to_text = patched_openai_speech_to_text_factory(model)
+    assert speech_to_text
     with temporary_file(
         "wav", tmp_file_directory=speech_to_text.tmp_file_directory
     ) as tmp: